create user authentication and file upload functionality

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,15 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -42,12 +51,35 @@ class PostsController extends Controller
 
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
-        ]);   
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
+        ]);
+
+        //handle the file upload
+        if($request->hasFile('cover_image')){
+
+            //get file name with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image 
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else {
+            $fileNameToStore = 'noimage.jpeg';
+        }   
 
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created successfully');
@@ -76,6 +108,10 @@ class PostsController extends Controller
     {
         $post = Post::find($id);
 
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -88,14 +124,36 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
+         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
-        ]);   
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
+        ]);
+
+        //handle the file upload
+        if($request->hasFile('cover_image')){
+
+            //get file name with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image 
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } 
 
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated successfully');
@@ -107,11 +165,26 @@ class PostsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            Storage::delete('public/cover_image/'.$post->cover_image);
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success','Post Deleted');
     }
+
+    //$post->onlyTrashed()->get();
+    // $allusers = User::withTrashed()->get(); include Soft delete
+    //$trashedusers = User::onlyTrashed()->get(); //Only soft deleted users
+
 }
